refactor(core): changesetSources -> changesetLoaders
`changesetSources` named the module's private internals — three source-shaped
loaders the caller never sees — while its only exports are `loadAppBootstrap`
and `orderDiffFiles`, and the entry point returns a bootstrap rather than a
changeset. "Source" also mostly means file contents here (`fileSource.ts`,
`sourceFetcher`, `sourceText`, `sourceBackedHighlight.ts`), so the name invited
a second reading of a word this tree already uses.

`changesetLoaders` keeps the accurate verb, says what gets loaded — the original
complaint about the bare `loaders.ts` — and covers the bootstrap entry point,
which is also a load.
